+++ eliot/__init__.py
@@ -1,5 +1,5 @@
 """
-Eliot: Logging as Storytelling.
+Eliot: Logging for Complex & Distributed Systems.
 """
 # Expose the public API:
 from ._message import Message
@@ -9,7 +9,7 @@ from ._validation import Field, fields, MessageType, ActionType
 from ._traceback import writeTraceback, writeFailure
 addDestination = Logger._destinations.add
 removeDestination = Logger._destinations.remove
-
+addGlobalFields = Logger._destinations.addGlobalFields
 
 # PEP 8 variants:
 start_action = startAction
@@ -18,18 +18,18 @@ write_traceback = writeTraceback
 write_failure = writeFailure
 add_destination = addDestination
 remove_destination = removeDestination
-
+add_global_fields = addGlobalFields
 
 
 __all__ = ["Message", "writeTraceback", "writeFailure",
            "startAction", "startTask", "Action",
            "Field", "fields", "MessageType", "ActionType",
            "ILogger", "Logger", "MemoryLogger", "addDestination",
-           "removeDestination",
+           "removeDestination", "addGlobalFields",
 
            # PEP 8 variants:
            "write_traceback", "write_failure", "start_action", "start_task",
-           "add_destination", "remove_destination",
+           "add_destination", "remove_destination", "add_global_fields",
            "to_file",
 
            "__version__",