+++ eliot/tests/test_message.py
@@ -14,7 +14,7 @@ except ImportError:
 
 from .._message import Message, _defaultAction
 from .._output import MemoryLogger
-from .._action import Action, startAction
+from .._action import Action, startAction, TaskLevel
 
 
 class MessageTests(TestCase):
@@ -119,15 +119,15 @@ class MessageTests(TestCase):
         L{Action} to the dictionary written to the logger.
         """
         logger = MemoryLogger()
-        action = Action(logger, "unique", "/", "sys:thename")
+        action = Action(logger, "unique", TaskLevel(level=[]),
+                        "sys:thename")
         msg = Message.new(key=2)
         msg.write(logger, action)
         written = logger.messages[0]
         del written["timestamp"]
         self.assertEqual(written,
                          {"task_uuid": "unique",
-                          "task_level": "/",
-                          "action_counter": 0,
+                          "task_level": [1],
                           "key": 2})
 
 
@@ -137,7 +137,8 @@ class MessageTests(TestCase):
         fields from the current execution context's L{Action} to the
         dictionary written to the logger.
         """
-        action = Action(MemoryLogger(), "unique", "/", "sys:thename")
+        action = Action(MemoryLogger(), "unique", TaskLevel(level=[]),
+                        "sys:thename")
         logger = MemoryLogger()
         msg = Message.new(key=2)
         with action:
@@ -146,8 +147,7 @@ class MessageTests(TestCase):
         del written["timestamp"]
         self.assertEqual(written,
                          {"task_uuid": "unique",
-                          "task_level": "/",
-                          "action_counter": 0,
+                          "task_level": [1],
                           "key": 2})
 
 
@@ -157,26 +157,27 @@ class MessageTests(TestCase):
         L{Action}, the process-specific global L{Action} is used.
 
         This ensures all messages have a unique identity, as specified by
-        task_uuid/task_level/action_counter.
+        task_uuid/task_level.
         """
         logger = MemoryLogger()
         msg = Message.new(key=2)
         msg.write(logger)
         written = logger.messages[0]
         del written["timestamp"]
+        nextLevel = _defaultAction._nextTaskLevel().level
+        prefix, suffix = nextLevel[:-1], nextLevel[-1]
+        expectedTaskLevel = prefix + [suffix - 1]
         self.assertEqual(written,
                          {"task_uuid":
-                              _defaultAction._identification["task_uuid"],
-                          "task_level": "/",
-                          "action_counter":
-                              _defaultAction._incrementMessageCounter() - 1,
+                          _defaultAction._identification["task_uuid"],
+                          "task_level": expectedTaskLevel,
                           "key": 2})
 
 
     def test_actionCounter(self):
         """
         Each message written within the context of an L{Action} gets its
-        C{action_counter} field incremented.
+        C{task_level} field incremented.
         """
         logger = MemoryLogger()
         msg = Message.new(key=2)
@@ -185,8 +186,8 @@ class MessageTests(TestCase):
                 msg.write(logger)
         # We expect 6 messages: start action, 4 standalone messages, finish
         # action:
-        self.assertEqual([m["action_counter"] for m in logger.messages],
-                         list(range(6)))
+        self.assertEqual([m["task_level"] for m in logger.messages],
+                         [[1], [2], [3], [4], [5], [6]])
 
 
     def test_writePassesSerializer(self):